refactor!: simplify proxy surface with credential auto-discovery
A ground-up cleanup that takes submux from a 110-file framework with many
dormant subsystems down to a focused, auto-discovering proxy for two
upstreams: Anthropic Claude Max and ChatGPT Codex.

What is gone

- src/router/ (12 files) — strategies, retry, failover. Routes pick from
  the pool directly via pool.by_provider.
- src/coordination/ (4 files) — multi-replica trait + Redis stub were
  never wired into the request path.
- src/storage/ (7 files) — Postgres account store + XChaCha20Poly1305
  sealer. Credentials now come from CLI auto-discovery or the TOML file.
- src/server/routes/admin.rs and the rate_limit / trace middleware stubs.
- src/telemetry/events.rs — EventBus had no subscribers. Routes call
  metrics::record_* directly.
- src/accounts/health.rs and the dead SubmuxConfig schema in config/.
- core::ProviderAdapter trait + execute / refresh_session / health_check
  stubs.
- All four CLI subcommands (config path/show/init). \`submux\` is the only
  verb now; edit the TOML to rotate the api_key.

What is new

- Credential auto-discovery (src/accounts/discovery/):
    Anthropic — macOS keychain item \"Claude Code-credentials\" (raw or
                hex-encoded JSON, both formats supported), with
                ~/.claude/.credentials.json as the Linux fallback.
    Codex    — ~/.codex/auth.json when auth_mode == \"chatgpt\".
  Resolution order per provider: env > config file > auto-discovery >
  none. The startup banner shows where each provider's token came from.
- TOML config file at \$XDG_CONFIG_HOME/submux/config.toml, written at
  first boot with mode 0600. Carries only api_key + optional overrides;
  tokens are auto-discovered.
- Optional inbound API key gate (src/server/middleware/auth.rs):
  constant-time bearer / x-api-key verification with a typed 401
  envelope. Disabled by default; /health, /ready, /metrics stay open
  even when enabled.
- Shared streaming relay state machine (src/streaming/relay.rs) with an
  SseTranslator trait. Both /v1/chat/completions and /codex/v1/messages
  share one unfold loop.
- Modern Rust 2018+ module layout: every \`foo/mod.rs\` replaced with a
  sibling \`foo.rs\`. Codified in CLAUDE.md as a structural rule.
- One concept per file: ApiKey, AccountId, ToolCallAccumulator,
  AnthropicToOpenAiRelay, TokenSource, FileConfig, Settings each live in
  their own file. No inline-defined struct/enum inside function bodies.

Renames for clarity

- AnthropicOAuthAdapter → AnthropicProxy
- OpenAiSubscriptionAdapter → CodexProxy
- providers/openai/ → providers/codex/
- chatgpt_session.rs → session.rs
- RuntimeConfig → Settings (config/resolve.rs → config/settings.rs)
- seed_from_runtime → seed_from_settings
- install_openai_adapter / openai_adapter() → codex::install / codex::current
- /healthz, /readyz → /health, /ready

Hygiene

- Zero \`as\` casts in production code. Every site uses From / TryFrom and
  branches explicitly on overflow.
- thiserror-only error enums; eyre::Result reserved for main.rs glue.
- Dead SubmuxError variants pruned (NoHealthyAccount, NeedsReauth,
  RetryExhausted, Translation, Storage, Config — none had callers).
- Cargo deps trimmed: drop sqlx + base64 (storage was their only
  consumer), drop the redis feature flag.
- All 133 unit tests pass; cargo fmt --check, cargo clippy --all-targets
  -D warnings clean.

Docs

- README rewritten around \`claude login\` + \`codex login\` + \`submux\`.
  The DevTools paste step is gone.
- docs/architecture.md rewritten from scratch to match current state.
- .github/README rewritten end to end.
- CLAUDE.md updated: new \"Project structure (modern Rust 2018+ layout)\"
  section codifying mod-free files, one-concept-per-file, method-first,
  minimal pub, no glob imports, ≤500 LOC per file, no narrowing \`as\`
  casts. Env var list reflects the new auto-discovery resolver. The
  Set-Cookie merge item is dropped from Outstanding Work (obsolete with
  auto-discovery); the sentinel handshake entry is tightened.

Verified end to end against this machine's actual credentials: both
providers auto-discovered, /v1/messages reaches api.anthropic.com,
/codex/v1/messages reaches chatgpt.com.

BREAKING CHANGE: removed \`submux config\` subcommands, the
SUBMUX_DATABASE_URL and SUBMUX_SEALER_KEY env vars, the /admin/* routes,
and renamed /healthz → /health, /readyz → /ready. The AnthropicProxy /
CodexProxy type names replace the old *Adapter names. Operators who set
api_key in the config file should re-confirm it after upgrade; everyone
else gains auto-discovery for free.

+++ src/providers/anthropic/quota.rs
@@ -9,12 +9,12 @@ use http::HeaderMap;
 use std::sync::Arc;
 use std::time::Duration;
 
+use crate::accounts::cooldown::{CooldownCache, CooldownReason};
 use crate::accounts::{quota::QuotaSnapshot, Account};
 use crate::constants::http_headers::{
     H_RL_5H_RESET, H_RL_5H_UTIL, H_RL_7D_RESET, H_RL_7D_UTIL, H_RL_FALLBACK, H_RL_STATUS,
 };
 use crate::constants::limits::{COOLDOWN_FALLBACK_SECS, COOLDOWN_UTIL_THRESHOLD};
-use crate::router::cooldown::{CooldownCache, CooldownReason};
 use crate::telemetry::metrics;
 
 fn parse_f32(headers: &HeaderMap, name: &str) -> Option<f32> {
@@ -77,10 +77,10 @@ pub async fn apply_quota(
     let account_id_str = account.id.to_string();
 
     if let Some(util) = snapshot.five_hour_util {
-        metrics::set_account_quota_utilization(&account_id_str, "5h", util as f64);
+        metrics::set_account_quota_utilization(&account_id_str, "5h", f64::from(util));
     }
     if let Some(util) = snapshot.seven_day_util {
-        metrics::set_account_quota_utilization(&account_id_str, "7d", util as f64);
+        metrics::set_account_quota_utilization(&account_id_str, "7d", f64::from(util));
     }
 
     if let (Some(util), Some(reset_at)) = (snapshot.five_hour_util, snapshot.five_hour_reset_at) {
@@ -88,7 +88,7 @@ pub async fn apply_quota(
             let delta = reset_at - Utc::now();
             let dur = delta
                 .to_std()
-                .unwrap_or_else(|_| Duration::from_secs(COOLDOWN_FALLBACK_SECS as u64));
+                .unwrap_or_else(|_| Duration::from_secs(COOLDOWN_FALLBACK_SECS));
             cooldown
                 .cool(account.id, CooldownReason::QuotaExhausted, dur, None)
                 .await;
